backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Cors==4.0.0
python-dotenv==1.0.0
gunicorn==21.2.0
pandas==2.0.3
redis==5.0.1
APScheduler==3.10.4
waitress==2.1.2
pyspark==3.5.0
Flask-Migrate==4.0.7